+++ scripts/tools/replay_demos.py
@@ -7,7 +7,12 @@
 
 """Launch Isaac Sim Simulator first."""
 
+
 import argparse
+
+# Import pinocchio in the main script to force the use of the dependencies installed by IsaacLab and not the one installed by Isaac Sim
+# pinocchio is required by the Pink IK controller
+import pinocchio  # noqa: F401
 
 from isaaclab.app import AppLauncher