+++ source/isaaclab_mimic/isaaclab_mimic/envs/__init__.py
@@ -12,6 +12,9 @@ from .franka_stack_ik_abs_mimic_env_cfg import FrankaCubeStackIKAbsMimicEnvCfg
 from .franka_stack_ik_rel_blueprint_mimic_env_cfg import FrankaCubeStackIKRelBlueprintMimicEnvCfg
 from .franka_stack_ik_rel_mimic_env import FrankaCubeStackIKRelMimicEnv
 from .franka_stack_ik_rel_mimic_env_cfg import FrankaCubeStackIKRelMimicEnvCfg
+from .franka_stack_ik_rel_visuomotor_mimic_env_cfg import FrankaCubeStackIKRelVisuomotorMimicEnvCfg
+from .pickplace_gr1t2_mimic_env import PickPlaceGR1T2MimicEnv
+from .pickplace_gr1t2_mimic_env_cfg import PickPlaceGR1T2MimicEnvCfg
 
 ##
 # Inverse Kinematics - Relative Pose Control
@@ -40,6 +43,24 @@ gym.register(
     entry_point="isaaclab_mimic.envs:FrankaCubeStackIKAbsMimicEnv",
     kwargs={
         "env_cfg_entry_point": franka_stack_ik_abs_mimic_env_cfg.FrankaCubeStackIKAbsMimicEnvCfg,
+    },
+    disable_env_checker=True,
+)
+
+gym.register(
+    id="Isaac-Stack-Cube-Franka-IK-Rel-Visuomotor-Mimic-v0",
+    entry_point="isaaclab_mimic.envs:FrankaCubeStackIKRelMimicEnv",
+    kwargs={
+        "env_cfg_entry_point": franka_stack_ik_rel_visuomotor_mimic_env_cfg.FrankaCubeStackIKRelVisuomotorMimicEnvCfg,
+    },
+    disable_env_checker=True,
+)
+
+gym.register(
+    id="Isaac-PickPlace-GR1T2-Abs-Mimic-v0",
+    entry_point="isaaclab_mimic.envs:PickPlaceGR1T2MimicEnv",
+    kwargs={
+        "env_cfg_entry_point": pickplace_gr1t2_mimic_env_cfg.PickPlaceGR1T2MimicEnvCfg,
     },
     disable_env_checker=True,
 )